+++ wallet/service/service.go
@@ -471,6 +471,7 @@ func NewService(log *zap.Logger, net *network.Network, apiV2 *jsonrpc.API, h Wal
 
 	// We don't use the `Service.handle() method to favour the improved logging
 	// inside the handler.
+	s.Handle(http.MethodGet, "/api/v2/methods", s.ListMethodsV2)
 	s.Handle(http.MethodPost, "/api/v2/requests", s.HandleRequestV2)
 
 	return s, nil